moved logout to actions/auth and used it in ProfileDropdown
