+++ actions/auth.ts
@@ -5,6 +5,8 @@ import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { Resend } from "resend";
 import { prisma } from "@/lib/db";
+import { ActionResult } from "@/lib/utils/types";
+import { lucia, validateRequest } from "@/lib/auth";
 
 async function checkUserExists(
   email: string
@@ -52,3 +54,23 @@ export const signInWithMagicLink = async (formData: FormData) => {
   console.log("email sent");
   redirect("/check-email");
 };
+
+export async function logout(): Promise<ActionResult> {
+  "use server";
+  const { session } = await validateRequest();
+  if (!session) {
+    return {
+      error: "Unauthorized",
+    };
+  }
+
+  await lucia.invalidateSession(session.id);
+
+  const sessionCookie = lucia.createBlankSessionCookie();
+  cookies().set(
+    sessionCookie.name,
+    sessionCookie.value,
+    sessionCookie.attributes
+  );
+  return redirect("/login");
+}
